app/pyproject.toml
[tool.poetry]
name = "timescaledb-backup"
version = "0.1.0"
description = ""
authors = ["Audun Ask Blaker <audun@askblaker.com>"]

[tool.poetry.dependencies]
python = "^3.8"
schedule = "^1.0.0"
minio = "^7.0.2"
psycopg2-binary = "^2.8.6"
pandas = "^1.2.3"
isoweek = "^1.3.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
isort = "^5.7.0"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
